Move Area to geo package
