+++ core/entities/src/commonMain/kotlin/geo/Area.kt
@@ -2,11 +2,7 @@
  * Copyright (c) 2024 Pointyware. Use of this software is governed by the GPL-3.0 license.
  */
 
-package org.pointyware.xyz.core.entities.ride
-
-import org.pointyware.xyz.core.entities.geo.LatLong
-import org.pointyware.xyz.core.entities.geo.Length
-import org.pointyware.xyz.core.entities.geo.LengthUnit
+package org.pointyware.xyz.core.entities.geo
 
 /**
  * Describes a geographical area.
